+++ app/SLCANSerial.hs
@@ -3,7 +3,7 @@ module Main where
 import Control.Monad.IO.Class
 import Data.Default.Class (Default(def))
 import System.Hardware.Serialport (CommSpeed(..), SerialPortSettings(..))
-import Network.CAN (MonadCAN)
+import Network.CAN (CANEndpoint)
 import Network.SLCAN (Transport(..))
 
 import qualified Control.Monad
@@ -29,12 +29,12 @@ main = do
     act
 
 act
-  :: ( MonadCAN m
-     , MonadIO m
-     )
-  => m ()
-act = do
+  :: MonadIO m
+  => CANEndpoint m
+  -> m ()
+act can = do
   Network.CAN.send
+    can
     $ Network.CAN.standardMessage
         -- vendorID SDO
         0x601
@@ -42,4 +42,5 @@ act = do
 
   Control.Monad.forever
     $ Network.CAN.recv
+        can
       >>= Control.Monad.IO.Class.liftIO . print