+++ src/Network/CAN.hs
@@ -1,7 +1,24 @@
 module Network.CAN
-  ( module Network.CAN.Class
+  ( CANEndpoint(..)
+  , send
+  , recv
   , module Network.CAN.Types
   ) where
 
-import Network.CAN.Class
 import Network.CAN.Types
+
+data CANEndpoint m = CANEndpoint
+  { canEndpointSend :: CANMessage -> m ()
+  , canEndpointRecv :: m CANMessage
+  }
+
+send
+  :: CANEndpoint m
+  -> CANMessage
+  -> m ()
+send = canEndpointSend
+
+recv
+  :: CANEndpoint m
+  -> m CANMessage
+recv = canEndpointRecv